Scorecard resource restructure (#8)
This restructures the `dx_scorecard` resource to support preserving
check history after reorders. The following list-based attributes have
changed to be maps:

- `levels`
- `check_groups`
- `checks`

Previously, the level and check group models both had a `key` attribute,
but now those are the **map keys** for the higher-level scorecard
attributes. Since checks is now a map as well, this requires the user to
begin defining keys for checks, to make their identities stable.

Our API expects each these three scorecard attributes to be lists, but
we store them as maps. So we have to convert back and forth. Here is how
that process works in `resource.go`:

- `modelToRequestBody`: Minimal change. We were previously looping over
the items in a list, but now we loop over the key-value entries in the
map.

- `responseBodyToModel`: This change is more complex. Our API responses
do not include the `key` of any of these objects, so it is up to us to
match them up. Fortunately, the API sorts the levels, check groups, and
checks deterministically based on their `ordering`/`rank`, and we can
run similar sorting in the provider to reliably match up an item in the
state with an item from an API response. This lets us know which pieces
of state to update.