Handle optional string fields (#13)
This updates several fields to be treated as optional like in the UI and
the core API:

- `check.description`
- `check.filter_message`
- `check.filter_sql`
- `check.external_url`

The API will default to the empty string `""` for these values, so
Terraform was previously getting confused about unexpected state
changes. But now we convert `""` to a `types.StringNull()` value in the
state, fixing the issue.