Sleep before deleting scorecards when running acceptance tests (#28)
This adds some logic to the scorecard API module, so if we are running
acceptance tests (ie. `TF_ACC=1`), then we will sleep for 2 seconds
before deleting.

This change is to avoid a race condition where a scorecard might be
asynchronously evaluating its checks while the delete request comes in,
causing 503 response and a leftover resource. We have not seen the issue
occur with normal production traffic, but the high-speed TF acceptance
tests are causing it to happen intermittently.